Refactor helper class
